removed tools and speech
