+++ setup.cfg
@@ -1,10 +1,10 @@
 [metadata]
 name = bia-bob
-version = attr: bia_bob.__version__
-description = A Jupyter-based assistant for working on bio-image analysis tasks
+version = attr: bia_tischi.__version__
+description = A Jupyter-based assistant for data analysis tasks
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/haesleinhuepf/bia-bob
+url = https://github.com/haesleinhuepf/bia-tischi
 author = Robert Haase
 author_email = robert.haase@tu-dresden.de
 license = BSD-3-Clause
@@ -34,15 +34,8 @@ install_requires =
     python-Levenshtein>=0.21.0
     scikit-image>=0.19.0
     langchain>=0.0.251
-    stackview>=0.6.3
-    napari-segment-blobs-and-things-with-membranes>=0.3.6
-    napari-skimage-regionprops
     pandas
     seaborn
-    SpeechRecognition
-    PyAudio
-    gTTS
-    pydub
 
 python_requires = >=3.8
 include_package_data = True
