--- src/bia_bob/_tests/__init__.py
@@ -1,2 +1,0 @@
-def test_nothing():
-    pass
