+++ src/bia_tischi/_tests/__init__.py
@@ -1,0 +1,2 @@
+def test_nothing():
+    pass
